Refactor: move RegisterNameConfiguration from presentation to project configuration package
