+++ src/main/java/uk/gov/register/presentation/representations/turtle/RecordListTurtleWriter.java
@@ -5,7 +5,7 @@ import org.apache.jena.rdf.model.ModelFactory;
 import uk.gov.register.configuration.RegisterNameConfiguration;
 import uk.gov.register.presentation.ItemConverter;
 import uk.gov.register.presentation.RegisterData;
-import uk.gov.register.presentation.config.RegisterDomainConfiguration;
+import uk.gov.register.configuration.RegisterDomainConfiguration;
 import uk.gov.register.presentation.representations.ExtraMediaType;
 import uk.gov.register.presentation.resource.RequestContext;
 import uk.gov.register.presentation.view.RecordListView;
