+++ src/main/java/uk/gov/register/presentation/representations/turtle/TurtleRepresentationWriter.java
@@ -3,7 +3,7 @@ package uk.gov.register.presentation.representations.turtle;
 import io.dropwizard.views.View;
 import org.apache.jena.rdf.model.Model;
 import uk.gov.register.configuration.RegisterNameConfiguration;
-import uk.gov.register.presentation.config.RegisterDomainConfiguration;
+import uk.gov.register.configuration.RegisterDomainConfiguration;
 import uk.gov.register.presentation.representations.RepresentationWriter;
 import uk.gov.register.presentation.resource.RequestContext;
