+++ src/main/java/uk/gov/register/presentation/view/BadRequestExceptionView.java
@@ -1,7 +1,7 @@
 package uk.gov.register.presentation.view;
 
 import uk.gov.register.presentation.RegisterData;
-import uk.gov.register.presentation.config.RegisterDomainConfiguration;
+import uk.gov.register.configuration.RegisterDomainConfiguration;
 import uk.gov.register.presentation.resource.RequestContext;
 import uk.gov.register.thymeleaf.ThymeleafView;
