+++ src/main/java/uk/gov/register/presentation/view/CsvRepresentationView.java
@@ -3,7 +3,7 @@ package uk.gov.register.presentation.view;
 import uk.gov.organisation.client.GovukOrganisation;
 import uk.gov.register.presentation.RegisterData;
 import uk.gov.register.presentation.config.PublicBody;
-import uk.gov.register.presentation.config.RegisterDomainConfiguration;
+import uk.gov.register.configuration.RegisterDomainConfiguration;
 import uk.gov.register.presentation.representations.CsvRepresentation;
 import uk.gov.register.presentation.resource.RequestContext;
