+++ src/main/java/uk/gov/register/presentation/view/EntryListView.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonValue;
 import uk.gov.organisation.client.GovukOrganisation;
 import uk.gov.register.presentation.RegisterData;
 import uk.gov.register.presentation.config.PublicBody;
-import uk.gov.register.presentation.config.RegisterDomainConfiguration;
+import uk.gov.register.configuration.RegisterDomainConfiguration;
 import uk.gov.register.presentation.dao.Entry;
 import uk.gov.register.presentation.representations.CsvRepresentation;
 import uk.gov.register.presentation.resource.IPagination;
