+++ src/main/java/uk/gov/register/presentation/view/RecordView.java
@@ -9,7 +9,7 @@ import uk.gov.register.presentation.ItemConverter;
 import uk.gov.register.presentation.FieldValue;
 import uk.gov.register.presentation.RegisterData;
 import uk.gov.register.presentation.config.PublicBody;
-import uk.gov.register.presentation.config.RegisterDomainConfiguration;
+import uk.gov.register.configuration.RegisterDomainConfiguration;
 import uk.gov.register.presentation.dao.Record;
 import uk.gov.register.presentation.representations.CsvRepresentation;
 import uk.gov.register.presentation.resource.RequestContext;
