+++ src/main/java/uk/gov/register/presentation/view/ViewFactory.java
@@ -7,7 +7,7 @@ import uk.gov.register.presentation.ItemConverter;
 import uk.gov.register.presentation.RegisterData;
 import uk.gov.register.presentation.config.PublicBodiesConfiguration;
 import uk.gov.register.presentation.config.PublicBody;
-import uk.gov.register.presentation.config.RegisterDomainConfiguration;
+import uk.gov.register.configuration.RegisterDomainConfiguration;
 import uk.gov.register.presentation.dao.Entry;
 import uk.gov.register.presentation.dao.Item;
 import uk.gov.register.presentation.dao.Record;
